feat: PC랜딩 세 번째 페이지 스켈레톤 적용
